params to interface
