+++ src/algorithm/blueprint/index.ts
@@ -1,5 +1,4 @@
 export * from './library';
-export * from './params';
 export * from './parser';
 export * from './signal';
 export * from './blueprint';
