+++ src/algorithm/blueprint/library.ts
@@ -289,6 +289,4 @@ export default {
    * - Conversion rate of anonymous sessions
    */
   NAVIGATION: Navigable_UX,
-  /** Alias for {@link User_Centric_Promoted_Content_Curations} algorithm. */
-  SIMPLE: User_Centric_Promoted_Content_Curations,
 } as const;
