+++ src/constants/index.ts
@@ -1,12 +1,4 @@
-import library from '../algorithm/blueprint/library';
-
-const presetAlgorithms: Record<Exclude<keyof typeof library, 'SIMPLE'>, string> = {
-  UNIQUE_JOURNEYS: 'Unique_Journeys'.toUpperCase(),
-  CONTENT_CURATION: 'User_Centric_Promoted_Content_Curations'.toUpperCase(),
-  AI_AGENTS: 'User_Intent_AI_Agents'.toUpperCase(),
-  RECOMMENDATIONS: 'Individually_Crafted_Recommendations'.toUpperCase(),
-  NAVIGATION: 'Navigable_UX'.toUpperCase(),
-};
+import {presetAlgorithms} from './algorithms';
 
 export default {
   /** Endpoint to get region. */
